link_bio/requirements.txt
reflex==0.7.1
mysql-connector-python
pykalman
scipy
fastapi
uvicorn
pandas
httpx